+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Deprecated
 ### Removed
 ### Fixed
+- Types for `req.error()` et al.
+- Types for `cds.error()`
 ### Security
 
 ## [0.15.0] - 2025-09-26